modificati tutti gli id da long a int, creati tutti i javadoc e la maggior parte dei commenti, cerato l'entità UserLibri con rispettivi repository e service

+++ src/main/java/com/tbr/lettura/model/Libro.java
@@ -10,7 +10,7 @@ import jakarta.persistence.Table;
 public class Libro {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long id;
+    private int id;
 
     private String title;
     private String author;
@@ -19,22 +19,144 @@ public class Libro {
     private String description;
 
     public Libro() {}
+    /*metodo che restituisce l'id del libro / 
+     * method that returns the id of the book
+     */
 
-    public Long getId() { return id; }
-    public void setId(Long id) { this.id = id; }
-
-    public String getTitle() { return title; }
-    public void setTitle(String title) { this.title = title; }
-
-    public String getAuthor() { return author; }
-    public void setAuthor(String author) { this.author = author; }
-
-    public String getGenre() { return genre; }
-    public void setGenre(String genre) { this.genre = genre; }
-
-    public Integer getYear() { return year; }
-    public void setYear(Integer year) { this.year = year; }
-
-    public String getDescription() { return description; }
-    public void setDescription(String description) { this.description = description; }
+    /**
+    * Retrieves the id of the book.
+    * 
+    * @return the id of the book.
+    */
+    public int getId() {
+         return id; 
+        }
+    /*metodo che modifica l'id del libro / 
+     * method that modifies the id of the book
+     */
+    
+    /**
+     * Modifies the id of the book.
+     * 
+     * @param id the new id of the book
+     */
+    public void setId(int id) {
+         this.id = id; 
+        }
+    
+    /*metodo che restituisce il titolo del libro / 
+     * method that returns the title of the book
+    */ 
+       
+    /**
+    * Retrieves the title of the book.
+    * 
+    * @return the title of the book.
+    */
+    public String getTitle() { 
+        return title; 
+    }
+    /*metodo che modifica il titolo del libro / 
+     * method that modifies the title of the book
+    */
+    
+    /**
+     * Modifies the title of the book.
+     * 
+     * @param title the new title of the book
+     */
+    public void setTitle(String title) { 
+        this.title = title; 
+    }
+    /*metodo che restituisce l'autore del libro / 
+     * method that returns the author of the book
+    */
+    
+    /**
+     * Retrieves the author of the book.
+     * 
+     * @return the author of the book
+     */
+    public String getAuthor() { 
+        return author; 
+    }
+    /*metodo che modifica l'autore del libro / 
+     * method that modifies the author of the book
+    */
+    /**
+     * Modifies the author of the book.
+     * 
+     * @param author the new author of the book
+     */
+    public void setAuthor(String author) { 
+        this.author = author; 
+    }
+    /*metodo che restituisce il genere del libro / 
+     * method that returns the genre of the book
+    */
+    /**
+     * Retrieves the genre of the book.
+     * 
+     * @return the genre of the book
+     */
+    public String getGenre() { 
+        return genre; 
+    }
+    /*metodo che modifica il genere del libro / 
+     * method that modifies the genre of the book
+    */
+    /**
+     * Modifies the genre of the book.
+     * 
+     * @param genre the new genre of the book
+     */
+    public void setGenre(String genre) { 
+        this.genre = genre; 
+    }
+    /*metodo che restituisce l'anno di pubblicazione del libro / 
+     * method that returns the year of publication of the book
+    */
+    
+    /**
+     * Retrieves the year of publication of the book.
+     * 
+     * @return the year of publication of the book
+     */
+    public Integer getYear() { 
+        return year; 
+    }
+    /*metodo che modifica l'anno di pubblicazione del libro / 
+     * method that modifies the year of publication of the book
+    */
+    /**
+     * Modifies the year of publication of the book.
+     * 
+     * @param year the new year of publication of the book
+     */
+    public void setYear(Integer year) { 
+        this.year = year; 
+    }
+    /*metodo che restituisce la descrizione del libro /
+     * method that returns the description of the book
+    */
+    
+    /**
+     * Retrieves the description of the book.
+     * 
+     * @return the description of the book
+     */
+    public String getDescription() { 
+        return description; 
+    }
+    /*metodo che modifica la descrizione del libro / 
+     * method that modifies the description of the book
+    */
+    /**
+     * Modifies the description of the book.
+     * 
+     * @param description the new description of the book
+     */
+    public void setDescription(String description) { 
+        this.description = description; 
+    }
 }
